pyproject.toml
[tool.poetry]
name = "delivery-prj"
version = "0.1.0"
description = ""
authors = ["vadim <144333283+vadyukkha@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11.11"
streamlit = "^1.40.2"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
